% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boycei.R
\name{boycei}
\alias{boycei}
\title{Boyce index model evaluation sub-function II}
\usage{
boycei(interval, obs, fit)
}
\description{
Not to be called directly by the user
}
\author{
'boycei' from the 'ecospat' package
}
