% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_env_strata.R
\name{create_envstrat}
\alias{create_envstrat}
\title{Create environmental strata for environmentally stratified pseudoabsence
sampling.}
\usage{
create_envstrat(
  env.stk,
  rAll = TRUE,
  save_it = TRUE,
  strat_dir = NA,
  poolsiz = 5 * 10^6,
  sampsiz = 4e+05,
  type
)
}
\arguments{
\item{env.stk}{raster stack with environmental layers of interest}

\item{rAll}{should raster be read entirely into memory? This is faster
but may blow the RAMs of your machine.}

\item{save_it}{Should a stratified sample of c. 400000 points be saved}

\item{strat_dir}{Directory to save the stratified sample in,
Not to be called directly by the user}
}
\description{
Create environmental strata for environmentally stratified pseudoabsence
sampling.
}
\author{
Philipp
}
