% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecospat_d2.R
\name{eco.adj.D2.glm}
\alias{eco.adj.D2.glm}
\title{Function originally written by A. Guisan (UNIL-ECOSPAT,Switzerland) and modified by C. Randin
(UNIL-ECOSPAT,Switzerland) S-Plus Function for calculating an adjusted D2 (see Weisberg, 1980) Takes any GLM
object as argument}
\usage{
eco.adj.D2.glm(glm.obj)
}
\description{
Not to be called directly by the user
}
\author{
'ecospat.adj.D2.glm' from the 'ecospat' package
}
