% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoBoyce.R
\name{ecoBoyce}
\alias{ecoBoyce}
\title{Boyce index model evaluation sub-function I}
\usage{
ecoBoyce(fit, obs, nclass = 0, window.w = "default", res = 100, PEplot = TRUE)
}
\description{
Not to be called directly by the user
}
\author{
'ecospat.boyce' from the 'ecospat' package
}
