% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_thres.R
\name{get_thres}
\alias{get_thres}
\title{Get threshold}
\usage{
get_thres(x, mean)
}
\arguments{
\item{x}{An object of class wsl.evaluation}

\item{mean}{Logical. If TRUE, a mean is applied to all thresholds}
}
\value{
numeric() or 'vector' with thresholds
}
\description{
Extracts thresholds from wsl.evaluation objects and names them so they can be fed
to the wsl.predict function. At the moment only averages over replicates can be obtained.
}
\examples{

# Take anguilla data set from dismo package
data("Anguilla_train")
vrs=c("SegSumT","USRainDays","USSlope")
env=Anguilla_train[,vrs]

### Check out wsl.glm
form.glm=as.formula(paste("Presence~",paste(paste0("poly(",vrs,",2)"),collapse="+")))

modi1=wsl.glm(pa=Anguilla_train$Angaus,
              env_vars = env,
              taxon="Angaus",
              replicatetype="cv",
              reps=5,
              project="prototest",
              mod_tag="test-glm",
              formula=form.glm,
              family="binomial",
              step=TRUE)

# Evaluate the model
eval1=wsl.evaluate.pa(modi1)

# Get thresholds
get_thres(eval1, mean = FALSE)
get_thres(eval1, mean = TRUE)

}
\author{
Philipp Brun, Yohann Chauvier
}
