% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_input_class.R
\docType{class}
\name{multi.input-class}
\alias{multi.input-class}
\alias{multi.input}
\title{An S4 class to store evaluation data}
\description{
An S4 class to store evaluation data
}
\section{Slots}{

\describe{
\item{\code{mod}}{Name of model algorithm to be called (character) e.g., 'glm'}

\item{\code{args}}{A list of arguments to be supplied to the model fitting alrogrithm}

\item{\code{tag}}{The name of the model setup (character)}

\item{\code{step}}{A logical indicating whether step function should be ran (for glm
and gam)}

\item{\code{weight}}{Should observations be weighted based on their prevalence?}
}}

\author{
Philipp Brun
}
