% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{multi}
\alias{multi}
\title{Define settings for function that should be supplied to wsl.flex}
\usage{
multi(mod, args, tag = "", step = FALSE, weight = FALSE)
}
\arguments{
\item{mod}{A character with the name of the function to be called. E.g. "gam"}

\item{args}{A list with arguments to be passed to the function specified in 'mod'}

\item{tag}{Character with name for model set-up}

\item{step}{Should step function be applied to update model}
}
\value{
Object of class 'multi'
}
\description{
Create a multi.input object that efficiently stores model specifications
}
\examples{

### Preliminary

data("Anguilla_train")
vrs=c("SegSumT","USRainDays","USSlope")

form.glm=as.formula(paste("Presence~",paste(paste0("poly(",vrs,",2)"),collapse="+")))
form.gbm=as.formula(Presence ~ .)
form.glm.2=as.formula(paste("Presence~",paste(vrs,collapse="+")))
feat=c("linear=true","quadratic=true","hinge=true","product=true","threshold=false")
form.gam=as.formula(paste("Presence~",paste(paste0("s(",vrs,")"),collapse="+")))

### Multi examples

multi("glm",list(formula=form.glm,family="binomial"),"glm-simple",step=TRUE)
multi("gbm",list(formula=form.gbm,
                 distribution = "bernoulli",
                 interaction.depth = 1,
                 shrinkage=.01,
                 n.trees = 3500),"gbm-simple")
multi("gam",list(formula=form.gam,family="binomial"),"gam-simple",step=FALSE)
multi("maxent",list(args=feat),"mxe-simple")
multi("randomForest",list(formula=form.gbm,ntree=500,maxnodes=NULL),"waud1")
multi("glm",list(formula=form.glm.2,family="binomial"),"glm-lin",step=TRUE)

}
\author{
Philipp Brun
}
