% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prd.R
\name{prd.pa}
\alias{prd.pa}
\title{Correctly feed the predict functions depending on model type (glm, gbm, maxent...)}
\usage{
prd.pa(mod, tst, clust = FALSE)
}
\description{
Not to be called directly by the user
}
\author{
Philipp Brun
}
