% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prd_pres.R
\name{prd.pres}
\alias{prd.pres}
\title{Correctly feed the predict functions depending on model type (ppm, glm, gbm, maxent...)}
\usage{
prd.pres(
  mod,
  env_vars,
  window,
  polly,
  meta,
  coefs,
  id.fact = NULL,
  valid.pres = NULL,
  env_samp = NULL
)
}
\description{
Not to be called directly by the user
}
\author{
Yohann Chauvier
}
