% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_sampling.R
\name{prop.sampling}
\alias{prop.sampling}
\title{Sample pseudo-absences proportional to target group distribution}
\usage{
prop.sampling(points, nsamples = 1000, res = 1, ...)
}
\arguments{
\item{points}{matrix or data.frame with column names 'x' and 'y' assumed to
be on the same scale and metric distances (m, km,...)}

\item{nsamples}{number of pseudoabsences to be generated}

\item{res}{resolution of the denstiy grid from which pseudo-absences are drawn.
default (1) corresponds to 1000 cells on the x-axis}

\item{...}{arguments passed on to the 'density' function from package spatstat.
Note in particluar the argument 'adjust' which controls the kernel size in the
density interpolation.}
}
\value{
nsamples x 2 matrix with drawn psuedo-absences
}
\description{
Uses the 'density' function from the package spatstat to create a density
surface of the supplied point pattern and samples pseudo-absences from this
density-distribution.
}
\examples{

### Load points

data(mypts)

### Test proportional sampling

pseu.1=prop.sampling(points=my.pts,nsamples=10000,adjust=1,res=1)
pseu.2=prop.sampling(points=my.pts,nsamples=10000,adjust=.5,res=1)
pseu.3=prop.sampling(points=my.pts,nsamples=10000,adjust=.2,res=1)
pseu.4=prop.sampling(points=my.pts,nsamples=10000,adjust=.1,res=1)

par(mfrow=c(2,2))

plot(pseu.1,pch=16,col="red",cex=.5)
points(my.pts)
plot(pseu.2,pch=16,col="red",cex=.5)
points(my.pts)
plot(pseu.3,pch=16,col="red",cex=.5)
points(my.pts)
plot(pseu.4,pch=16,col="red",cex=.5)
points(my.pts)

}
\author{
Philipp Brun
}
