% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{stratify}
\alias{stratify}
\title{Subsample based on two stratification critera}
\usage{
stratify(spp, type, sampsiz)
}
\arguments{
\item{spp}{Spatial data.frame}

\item{type}{either 'env.strat' or 'env.semi.strat'}

\item{sampsiz}{desired size of the subsample
Not to be called directly by the user}
}
\description{
Subsample based on two stratification critera
}
\references{
Descombes, P., Chauvier, Y., Brun, P., Righetti, D., Wüest, R. O., Karger, D. N., ... &
Zimmermann, N. E. (2022). Strategies for sampling pseudo-absences for species distribution
models in complex mountainous terrain. bioRxiv, 2022-03.
}
\author{
Philipp Brun
}
