% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsample_strategic.R
\name{upsample_strategic}
\alias{upsample_strategic}
\title{Sample a supsample from a large set of points with a minimum distance
constraint}
\usage{
upsample_strategic(spdf, lim_dist, n_tot, warnig = TRUE)
}
\arguments{
\item{spdf}{SpatialPoints or SpatialPointsDataFrame object}

\item{lim_dist}{Minimum tolerated distance between the points}

\item{n_tot}{Minimum tolerated distance between the points}
}
\value{
SpatialPoints or SpatialPointsDataFrame with at least
lim_dist between all points
}
\description{
Thin a spatial points object by number of points or minimum distance
}
\details{
iteratively samples points and rejects them if min thist
to all existing points i the sample are not respected
}
\references{
Descombes, P., Chauvier, Y., Brun, P., Righetti, D., Wüest, R. O., Karger, D. N., ... &
Zimmermann, N. E. (2022). Strategies for sampling pseudo-absences for species distribution
models in complex mountainous terrain. bioRxiv, 2022-03.
}
\author{
Philipp brun
}
