% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsample_thin.R
\name{upsample_thin}
\alias{upsample_thin}
\title{Sample a supsample from a large set of points with a minimum distance
constraint}
\usage{
upsample_thin(spdf, lim_dist, n_tot)
}
\arguments{
\item{spdf}{SpatialPoints or SpatialPointsDataFrame object}

\item{lim_dist}{Minimum tolerated distance between the points}

\item{n_tot}{Minimum tolerated distance between the points}
}
\value{
SpatialPoints or SpatialPointsDataFrame with at least
lim_dist between all points
}
\description{
Thin a spatial points object by number of points or minimum distance
}
\details{
iteratively samples points and rejects them if min thist
to all existing points i the sample are not respected
}
\author{
Philipp
}
