% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permut_importance.R
\name{var.imp}
\alias{var.imp}
\title{Assess predictor importance using random permutations}
\usage{
var.imp(model, data, nperm, type, rescale)
}
\arguments{
\item{model}{A model output. e.g. GLM, GAM, GBM, random forest, glmnet outputs...}

\item{data}{A data.frame of environmental values.}

\item{nperm}{How many random permutations should be applied per predictor?}

\item{type}{Response type. Type of desired response.}

\item{rescale}{Should the results be rescaled from a scale from 1 to 100?}
}
\value{
Average model importance score of each predictor.
}
\description{
This function is based on the principle of permutation importance. This metric is applicable
to any type of model. The basic idea is to consider a variable important if it has a positive
effect on the prediction accuracy (classification), or MSE (regression). The risk of using this
metric is a potential bias towards collinear predictive variables. It is important to note that
this function is fully compatible with singular SDM output of the package.
}
\examples{

# Take anguilla data set from dismo package
data("Anguilla_train")
vrs = c("SegSumT","USRainDays","USSlope")

# Apply a simple GLM and measure each predictors' contribution to the model
form.glm = as.formula(paste("Angaus~",paste(paste0("poly(",vrs,",2)"),collapse="+")))
glm.calib = glm(form.glm,data=Anguilla_train,family="binomial")
imp.test = var.imp(model = glm.calib,
                   data = Anguilla_train[,vrs],
                   nperm = 10,
                   type = "response",
                   rescale = TRUE);imp.test

# Apply a glmnet model and measure each predictors' contribution to the model

}
\author{
Patrice Descombes
}
