% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsl_evaluation_class.R
\docType{class}
\name{wsl.evaluation-class}
\alias{wsl.evaluation-class}
\alias{wsl.evaluation}
\title{An S4 class to store evaluation data}
\description{
An S4 class to store evaluation data
}
\section{Slots}{

\describe{
\item{\code{meta}}{A list with meta information}

\item{\code{thres}}{A vector with externally supplied thresholds}

\item{\code{performance}}{A list with model performance estimates}
}}

\author{
Philipp Brun
}
