% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsl_obs_filter.R
\name{wsl.obs.filter}
\alias{wsl.obs.filter}
\title{Filter set of observations}
\usage{
wsl.obs.filter(o.xy, a.xy = NULL, grid)
}
\arguments{
\item{o.xy}{Object of class 'matrix' or 'data frame' with two columns named
"x" and "y". Must be used alone to apply grid filetring to one set of observations
(presences or absences).}

\item{a.xy}{Object of class 'matrix' or 'data frame' with two columns named "x" and
"y". NULL by default. If used, "o.xy" becomes presences and "a.xy" absences.
Resolution grid filtering is applied to both observation sets, but coordinates of
"a.xy" falling within the same grid cells as "o.xy" are removed.}

\item{grid}{Object of class 'RasterLayer', 'RasterBrick' or 'RasterStack' of
desired resolution and extent.}
}
\value{
Object of class 'matrix' or 'data frame' with two columns named "x" and "y"
comprising the new set of observations filtered at grid resolution. If "a.xy" is
used the output is a list of two data.frame: filtered presences and filtered absences
}
\description{
Filter presences / absences through a chosen resolution grid.
}
\examples{

### Load my binary observations species data

library(raster)

data(var_select_XYtest)
data(exrst)

### wsl.obs.filter(): example for the first species

   # Loading observations: presences and absences

presences = coordinates(mySP[[1]])[myPA[[1]] \%in\% "1",]
absences = coordinates(mySP[[1]])[myPA[[1]] \%in\% "0",]

   # Loading grid

r.layer = rst[[1]]

   # To filter observations by the grid only

pres.filtered = wsl.obs.filter(presences,grid=r.layer)
abs.filtered = wsl.obs.filter(absences,grid=r.layer)

   # To filter observations by the grid & remove abs. in cells where we also find pres.

PresAbs.filtered = wsl.obs.filter(presences,absences,r.layer)

   # Count presences (same filtering)

nrow(PresAbs.filtered[[1]])
nrow(pres.filtered)

   # Count Absences (filtering plus removal of duplicated absences)

nrow(abs.filtered)
nrow(PresAbs.filtered[[2]])

   # Visual

par(mfrow=c(1,2))
plot(presences)
plot(pres.filtered)

par(mfrow=c(1,2))
plot(absences)
plot(abs.filtered)

}
\author{
Yohann Chauvier
}
