% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsl_ppm_env.R
\name{wsl.ppm.env}
\alias{wsl.ppm.env}
\title{Poisson Point Process Models (PPPM) 'env'}
\usage{
wsl.ppm.env(env_vars, mask)
}
\arguments{
\item{env_vars}{Object of class 'RasterBrick' or 'RasterStack'. Same extent and resolution as 'mask'}

\item{mask}{Object of class 'RasterLayer' or 'RasterBrick' or 'RasterStack'. Mask is defined
according to non NAs values and the resolution of the study}
}
\value{
List of objects of class 'im'
}
\description{
Convert envrionmental data for 'wsl.ppmO' fitting.
}
\examples{

### Load

data(exrst)
data(xy_ppm)
mypoints = xy.ppm[,c("x","y")]

### Define mask

maskR = mask(rst[[1]],shp.lonlat)

### Define your environments

   # For 'wsl.ppmGlasso' (observations focus)
envG = raster::extract(rst,mypoints)

   # For 'wsl.ppmO' (study area focus)
envO = wsl.ppm.env(rst,maskR)

}
\author{
Yohann Chauvier
}
