% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsl_ppm_qscheme.R
\name{wsl.ppm.qscheme}
\alias{wsl.ppm.qscheme}
\title{To prepare PPM general quadratic environment with input species points}
\usage{
wsl.ppm.qscheme(data, area.win, quads)
}
\description{
Not to be called directly by the user
}
\author{
Yohann Chauvier
}
