% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsl_ppm_window.R
\name{wsl.ppm.window}
\alias{wsl.ppm.window}
\title{Poisson Point Process Models (PPPM) 'window'}
\usage{
wsl.ppm.window(mask, val = 1, owin)
}
\arguments{
\item{mask}{Object of class 'RasterLayer' or 'RasterBrick' or 'RasterStack'. Mask is defined
according to non NAs values. Defined study area in pixels. Same as in 'wsl.quadrature'}

\item{val}{Values that are assigned to the 'owin' or 'im' mask}

\item{owin}{TRUE or FALSE}
}
\value{
Object of class 'owin' or 'im'
}
\description{
Set up a study window of class 'owin' or 'im' necessary to run 'wsl.ppmO' fit function.
}
\examples{

### Load

data(AlpineConvention_lonlat)
data(exrst)

### Define mask

maskR = mask(rst[[1]],shp.lonlat)

### Run function

wind = wsl.ppm.window(mask = maskR,
                      val = 1,
                      owin = TRUE)

}
\author{
Yohann Chauvier
}
