% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_predict.R
\name{wsl.predictGlasso}
\alias{wsl.predictGlasso}
\title{SDM predict function for glmnet outputs}
\usage{
wsl.predictGlasso(mod, predat)
}
\arguments{
\item{mod}{A glmnet model output.}

\item{predat}{Same spatial layers used in the fitted model or an object of class 'data.frame'
or  matrix' defining a sample of the target layers by keeping same order for columns.
If spatial layers, when 'wsl.ppmGlasso' used, object of class 'RasterStack' or RasterBrick'.}
}
\value{
RasterLayer or Vector.
}
\description{
This function can make predictions based on any type of glmnet models, i.e.
even complex model calibrated with polynomial terms. Additionally, the predict
function avoids errors when pedictions are made based on predictors whose
values equal all one single value (very useful to apply bias covariate correction).
}
\examples{

}
\author{
Yohann Chauvier
}
