% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsl_prediction_class.R
\docType{class}
\name{wsl.prediction-class}
\alias{wsl.prediction-class}
\alias{wsl.prediction}
\title{An S4 class to store prediction data}
\description{
An S4 class to store prediction data
}
\section{Slots}{

\describe{
\item{\code{meta}}{A list with meta information}

\item{\code{thres}}{A vector with externally supplied thresholds}

\item{\code{predictions}}{A list with model predictions}
}}

\author{
Philipp Brun
}
