% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsl_pseudoabsences_class.R
\docType{class}
\name{wsl.pseudoabsences-class}
\alias{wsl.pseudoabsences-class}
\alias{wsl.pseudoabsences}
\title{An S4 class to store pseudoabsence data}
\description{
An S4 class to store pseudoabsence data
}
\section{Slots}{

\describe{
\item{\code{meta}}{A list with meta information}

\item{\code{type}}{The type of pseudoabsence sampling}

\item{\code{pa}}{a vector with presences and pseudo absences}

\item{\code{env_vars}}{data.frame with environmental predictors}

\item{\code{xy}}{a matrix with coordinates of the points}
}}

\author{
Philipp Brun
}
