% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsl_quads_class.R
\docType{class}
\name{wsl.quads-class}
\alias{wsl.quads-class}
\alias{wsl.quads}
\title{An S4 class to store quadrature  objects}
\description{
An S4 class to store quadrature  objects
}
\section{Slots}{

\describe{
\item{\code{coordinates}}{informations on the generated quadratures}

\item{\code{predictors}}{values extracted for each quadrature}
}}

\author{
Yohann Chauvier, Philipp Brun
}
