% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsl_varKeep.R
\name{wsl.varKeep}
\alias{wsl.varKeep}
\title{Pre-select set of predictors}
\usage{
wsl.varKeep(
  PPower.object,
  ras,
  corTEST = 0.7,
  vifTEST = FALSE,
  corVIF = 0.7,
  ...
)
}
\arguments{
\item{PPower.object}{'wsl.varPPower' object}

\item{ras}{Object of class 'RasterLayer' or 'list' of 'RasterLayer'. Same as in wsl.varPPower().
Layers must be of same resolution i.e. excatly the same number of cells}

\item{corTEST}{Numeric. Threshold of accepted correlation applied in the descending tests}

\item{vifTEST}{Logical. If TRUE, a VIF test is applied to the set of predictors that succeeded
descending correlation tests}

\item{corVIF}{Numeric. Threshold of accepted correlation within VIF if "vifTEST=TRUE"}

\item{...}{Arguments passed on to the cor() function used for "corTEST"}
}
\value{
Object of class list with three elements: 1) ranking of non correlated predictors,
2) ranking of leftovers correlated predictors, 3) vif test results
}
\description{
Rank spatial predictors "ras" in descending order following their mean predictive
power found with wsl.varPPower(), and apply descending successive correlations tests
to obtain a pre-selection of predictors. Additionally a Variance Inflation Factor
Test (VIF) may also be applied to predictors which succeeded the descending tests to
identify the ones that show high mutlicollinearity.
}
\examples{
### Data Preparation

# Load environmental rasters and assign random raster Class
data(exrst)
rCLASS = c(rep("Pollution",4),rep("Climate",4),rep("LandCover",4))

# Create a list of rasters out of the rasterBRICK
rasterL = unstack(rst)

# Load my binary observations species data
data(var_select_XYtest.RData)

### wsl.varPPower(): example with a data.frame & rasters not in classes

PPower.DF = wsl.varPPower(points=sp.DF[,c("x","y")],
                          val = sp.DF$myPA,
                          species = sp.DF$spCODES,
                          ras = rasterL,
                          rasCLASS = NULL,
                          mlinear = FALSE,
                          glmMODE = "binomial",
                          weight = sp.DF$myWEIGHT,
                          poly = TRUE,
                          polyBRUT = TRUE,
                          parallel = FALSE,
                          cores = NULL)


### wsl.varPPower(): example with SpatialPointsDataFrame & rasters in classes

PPower.DF = wsl.varPPower(points=sp.DF[,c("x","y")],
                          val = sp.DF$myPA,
                          species = sp.DF$spCODES,
                          ras = rasterL,
                          rasCLASS = NULL,
                          mlinear = FALSE,
                          glmMODE = "binomial",
                          weight = sp.DF$myWEIGHT,
                          poly = TRUE,
                          polyBRUT = TRUE,
                          parallel = FALSE,
                          cores = NULL)

### wsl.varPPower(): example with SpatialPointsDataFrame & rasters in classes

# Checking length of every species input
c(length(mySP),length(myPA),length(spCODES),length(myWEIGHT))

# Running the function
PPower.SPDF = wsl.varPPower(points = mySP,
                            val = myPA,
                            species = spCODES,
                            ras = rasterL,
                            rasCLASS = rCLASS,
                            mlinear = FALSE,
                            glmMODE = "binomial",
                            weight = myWEIGHT,
                            poly = TRUE,
                            polyBRUT = TRUE,
                            parallel = FALSE,
                            cores = NULL)


 ### wsl.varKeep() with a descending correlation of 0.7 with a VIF test at 0.7

# Example with no raster classes
KeepVar = wsl.varKeep(PPower.object = PPower.DF,
                      ras = rasterL,
                      corTEST = 0.7,
                      vifTEST = TRUE,
                      corVIF = 0.7,
                      use="complete.obs")

#Example with raster classes
KeepVar = wsl.varKeep(PPower.object = PPower.SPDF,
                      ras = rasterL,
                      corTEST = 0.7,
                      vifTEST = TRUE,
                      corVIF = 0.7,
                      use = "complete.obs")

}
\author{
Yohann Chauvier
}
