% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_env.R
\name{get_env}
\alias{get_env}
\title{Extract environments based on yearly observations (local or URL sources)}
\usage{
get_env(
  env_layers = NULL,
  env_id = NULL,
  year_id = NULL,
  obs = data.frame(),
  d.path = getwd()
)
}
\arguments{
\item{env_layers}{Vector of file names or URL links of desired environmental layers. The
given strings must have the identity of the variable and the year informed.}

\item{env_id}{Vector. What are the unique names (as specified in 'env_layers') of the target
variables?}

\item{year_id}{Vector. What are the unique years (as specified in 'env_layers') we want the
environment from?}

\item{obs}{A data.frame with three columns (1-Longitude, 2-Latitude and 3-year of sampling).
The observations needs to be from the same CRS as the 'env_layers'.}

\item{d.path}{Character. The folder path where the environmental layers should be saved or stored.}
}
\value{
A data.frame describing for each observations their associated environmental values.
}
\description{
Based on local files or URL sources, this function extracts environmental
values from given observations at their year of sampling. Although the function
was developped mainly for CHELSA (https://chelsa-climate.org/), it can easily be
used for other type of data sources if following the same file format (i.e. single
raster files, informed environments and year in the URL or file names)
}
\examples{

# Packages
library(wsl.biodiv)
data(xy_ppm)

# URL links
url.base = "https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V1/chelsa_cruts/"
url.sprint1 = sprintf(paste0("tmax/CHELSAcruts_tmax_10_\%d_V.1.0.tif"),1901:1911)
url.sprint2 = sprintf(paste0("tmin/CHELSAcruts_tmin_10_\%d_V.1.0.tif"),1906:1921)
URL.links = paste0(url.base,c(url.sprint1,url.sprint2))
 
# Dummy observations
random.year = (1901:1921)[sample(1:21,nrow(xy.ppm),replace=TRUE)]
obs = data.frame(xy.ppm[,c("x","y")],year=random.year)

# Run function
sp.env = get_env(env_layers = URL.links,
                 env_id = c("tmax","tmin"),
                 year_id = 1901:1921,
                 obs = obs,
                 d.path = getwd())

}
\references{
...
}
