% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thin_them_function.R
\name{thin_them}
\alias{thin_them}
\title{Thin a spatial points object by number of points or minimum distance}
\usage{
thin_them(spdf, lim_dist = NA, lim_n = NA)
}
\arguments{
\item{spdf}{SpatialPoints or SpatialPointsDataFrame object}

\item{lim_dist}{Minimum tolerated distance between the points}

\item{lim_n}{Number of points to be kept}
}
\value{
thinned SpatialPoints or SpatialPointsDataFrame object
}
\description{
Thin a spatial points object by number of points or minimum distance
}
\details{
iteratively the closest point pairs are compared and the
point with the lower overall distinctiveness is removed
}
\author{
Philipp
}
